GET-6387 Migrate GCP configurations to public repo (#1)
# Overview

Migrates the GCP setup instructions, scripts, and Terraform module into
a standalone repository. The intent is for this repository to become
public-facing. Clients will be directed from Metronome's documentation
to use the tools in this repository to prepare their GCP project for
Metronome-managed invoicing. The configurations provide several possible
methods to deploy a Workload Identity Federation Pool and a Google
Service Account to their project with the appropriate configuration.

These configs have been reviewed and merged already to the [marketplaces
repository](https://github.com/Metronome-Industries/marketplaces/tree/main/terraform/metronome-gcp-marketplace-integration)
while we were testing and debugging things internally. This PR just
copies the configurations over to a public repository so that clients
may access them.

Note that this repository is not yet public, I'll convert it to public
once we are ready to push the GCP integration GA.

Co-authored-by: Jon Fosdick <jon.fosdick@aprime.io>