GET-6937 rm "coming soon" warning from gcp terraform repo (#3)
Makes two docs changes:
* removes coming soon banner
* removes an irrelevant command from one of the debugging steps


RE: irrelevant command:
Deleting a "verification" step from the gcloud setup docs which
encourages the user to check the permissions on the service account
these configs deploy. This instruction is a holdover from when we had
these configs actually attaching permissions to the service account, a
plan we abandoned after realizing that some of the permissions can only
be granted through the UI. If the user has followed instructions exactly
through getting to this part of the docs, we actually expect the
response to be empty (they will be separately instructed later to go
into the GCP console and add the permissions manually), so I think
leaving this here will just cause confusion.

Co-authored-by: Jon Fosdick <jon.fosdick@aprime.io>